Writing cookies: Prefix wildcard domains with "."
If a cookie matches a wildcard domain, e.g. "(anything).example.com",
then the line written out to a Netscape-formatted cookies.txt file must
begin with a dot, e.g. ".example.com" not "example.com".  This is
redundant with the second field of the line, a TRUE/FALSE value
indicating whether the domain is a wildcard, but that's the way it goes.