Prevent malformed WWW-Authenticate headers from causing infinite loops
The `Mechanize::HTTP::WWWAuthenticateParser` parser gets stuck in a loop
in some cases when the WWW-Authenticate header is malformed.

The string scanner will continue to loop if it has not reached the end
of the header string AND if the remainder of the string does not contain a
recognisable auth-scheme.

This PR changes the behaviour such that it  will break out of the loop
if the remaining header string is not recognisable (rather than getting
stuck in a loop).